Исправляем тип Accrual

+++ internal/model/model.go
@@ -21,7 +21,7 @@ type Order struct {
 	UserID     int64
 	Number     string
 	Status     string
-	Accrual    int
+	Accrual    float32
 	UploadedAt time.Time
 }
